+++ src/gort/observer.py
@@ -11,10 +11,11 @@ from __future__ import annotations
 import asyncio
 import logging
 import re
+import signal
 from contextlib import contextmanager
 from time import time
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Callable
 
 import numpy
 import pandas
@@ -27,6 +28,7 @@ from gort.tile import Coordinates
 from gort.tools import (
     build_guider_reply_list,
     cancel_task,
+    handle_signals,
     insert_to_database,
     register_observation,
 )
@@ -41,6 +43,24 @@ if TYPE_CHECKING:
 __all__ = ["GortObserver"]
 
 
+class InterrupHandlerHelper:
+    """Helper for handling interrupts"""
+
+    def __init__(self):
+        self._callback: Callable | None = None
+
+    def run_callback(self):
+        if self._callback is not None:
+            self._callback()
+
+    def set_callback(self, cb: Callable | None):
+        self._callback = cb
+
+
+interrupt_helper = InterrupHandlerHelper()
+interrupt_signals = [signal.SIGINT, signal.SIGTERM]
+
+
 class GortObserver:
     """A class to handle tile observations.
 
@@ -52,10 +72,18 @@ class GortObserver:
         The :obj:`.Tile` with the information about the observation.
     mask_positions_pattern
         The ``spec`` fibre mask positions to use.
+    on_interrupt
+        Callback to be called when the observation is interrupted.
 
     """
 
-    def __init__(self, gort: Gort, tile: Tile, mask_positions_pattern: str = "P1-*"):
+    def __init__(
+        self,
+        gort: Gort,
+        tile: Tile,
+        mask_positions_pattern: str = "P1-*",
+        on_interrupt: Callable | None = None,
+    ):
         self.gort = gort
         self.tile = tile
 
@@ -71,6 +99,8 @@ class GortObserver:
 
         self.overheads: dict[str, tuple[float, float]] = {}
 
+        interrupt_helper.set_callback(on_interrupt)
+
     def __repr__(self):
         return f"<GortObserver (tile_id={self.tile.tile_id})>"
 
@@ -80,6 +110,7 @@ class GortObserver:
 
         return len(self.standards.standards) > 0
 
+    @handle_signals(interrupt_signals, interrupt_helper.run_callback)
     async def slew(self):
         """Slew to the telescope fields."""
 
@@ -143,6 +174,7 @@ class GortObserver:
         with self.register_overhead("slew:slew"):
             await asyncio.gather(*cotasks)
 
+    @handle_signals(interrupt_signals, interrupt_helper.run_callback)
     async def acquire(self, guide_tolerance: float | None = None, timeout: float = 180):
         """Acquires the field in all the telescopes. Blocks until then.
 
@@ -267,6 +299,7 @@ class GortObserver:
 
         self.write_to_log("All telescopes are now guiding.")
 
+    @handle_signals(interrupt_signals, interrupt_helper.run_callback)
     async def expose(
         self,
         exposure_time: float = 900.0,
@@ -376,6 +409,7 @@ class GortObserver:
         else:
             return exposures
 
+    @handle_signals(interrupt_signals, interrupt_helper.run_callback)
     async def finish_observation(self):
         """Finishes the observation, stops the guiders, etc."""